dev-requirements.txt
pytest-ckan
pytest-mock
pytest-factoryboy
pytest==4.6.5
black
